backend/pyproject.toml
[project]
name = "github-semantic-search"
version = "0.1.0"
description = "A semantic search engine for GitHub repositories"
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
keywords = ["github", "semantic-search", "embeddings", "ai", "search"]
authors = [{ name = "Project Author", email = "author@example.com" }]
dependencies = [
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "pydantic>=2.6.3",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.1",
    "pydantic-settings>=2.8.1",
    "openai>=1.66.3",
    "google-genai>=1.5.0",
    "aiohttp>=3.9.3",
    "tenacity>=8.2.3",
    "sqlalchemy>=2.0.28",       # For database ORM
    "alembic>=1.13.1",          # For database migrations
    "asyncpg>=0.29.0",          # For async PostgreSQL support
    "structlog>=24.1.0",        # For structured logging
]

[project.optional-dependencies]
dev = ["mypy>=1.8.0", "ruff>=0.3.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.hatch.build.targets.wheel]
packages = ["src/github_semantic_search"]

[tool.uv]
package = true


[dependency-groups]
dev = ["ruff>=0.11.0"]
